Use close_cb to be informed when channel is closed
Checking if the process has exited in the output handler is unreliable
since the process could still be running shortly after it finishes
outputting and the output handler is called for the last time. In that
case the code will never be informed that the process has finished.

Instead, we use the close_cb callback to listen for when the pipe is
closed (i.e. process has finished executing).